replace ArduSubPilot with Shuttle
