--- devices/ArduSubPilot/ardusubpilot/__init__.py
@@ -1,3 +1,0 @@
-__version__ = '0.1.0'
-
-from ardusubpilot.device_controller import DeviceController
