+++ devices/Shuttle/shuttle/__init__.py
@@ -1,0 +1,3 @@
+__version__ = '0.1.0'
+
+from shuttle import shuttle_connector, websocket_connector
