+++ devices/Shuttle/shuttle/config.py
@@ -7,4 +7,4 @@ load_dotenv()
 # String used by pymavlink to connect with the ROV or simulator.
 # Needs to be changed depending on the physical setup. Documentation: 
 # http://mavlink.io/en/mavgen_python/
-CONNECTION_STRING = 'udpin:0.0.0.0:14550'   
+CONNECTION_STRING = 'udpin:0.0.0.0:14550'
